refactor: consolidate type system & eliminate redundant type definitions
